Stage 6: the integer and the array
Palette.glsl was the one shader in the corpus that would not convert, and
it named five gaps doing it: an array type, a subscript, an integer
value, and the mask that keeps the index in range. That is what one gap
looks like when it is genuinely four things, and all four are closed in
eacp now - Int, the operators only integers have, and a constant array to
subscript with one - so this stage is mostly about what the transpiler
does with them.

The README had multi-buffer down as stage 6 and the integer type as the
other thing the table was pointing at. The table decided: multi-buffer is
not measured by anything in the corpus yet, and the integer was the only
row there. Multi-buffer becomes stage 7.

The parser takes array declarations and the constructor that fills one -
vec3 palette[4] = vec3[4](...) - dropping the size on both sides, since
the element count already carries it and two sources for one number is
one more than the port can use. It also gains the shift level, which was
missing from its precedence table outright: << and >> lexed and then had
nowhere to go.

Lowering keeps substituting an integer local that folds to a constant,
which is what keeps an ordinary loop counter out of the port entirely.
What changed is the answer when it does not fold: it is an int now rather
than a diagnostic. A uint still is not - the EDSL's unsigned type is the
compute thread id and the buffer index computed from it, and nothing in a
fragment shader reaches it. A constant array survives lowering as itself:
there is nothing to fold it into.

The counter of a loop the port keeps is an int too, where stage 5 made it
a float. That was not a preference; it stopped type-checking. Both sides
of a loop header have to agree, and a bound spelled int(iTime) is now
genuinely an integer, so a float counter beside it would be comparing
across the two vocabularies - which GLSL will not do either. Counting in
the type the shader wrote is what keeps the rest of the body consistent,
and every crossing is one the source already wrote out.

iFrame goes the same way for the same reason. It is an int on the
Shadertoy page and was a float here only for as long as there was nothing
else for it to be. It is a Uniform<Int> now, which leaves iDate as the
one deviation from the real uniform set - and that is a clock this
runtime does not have rather than a type the EDSL is missing.

The emitter grew Type::Int and Type::Array, int()/float() as the two
crossings rather than a report and an identity, the integer operators
including ~, an array declaration as array(e0, e1, ...) that the wrapping
layout can break like any other call, and the subscript itself.

One thing there is worth recording, because it is not where you would
look for it. A literal has no type of its own: index & 3 needs a 3 and
the int(uv.x * 4.0) it was built from still needs its 4.0f, and the same
AST node kind carries both. So which one a number is spelled as is
decided by where it sits, walked once per statement immediately before
that statement is emitted - the types the walk reads are the ones the
statements above it left behind. The result is a set of nodes rather than
a flag threaded through emission, which is what lets the line-wrapping
path reach the same answer without knowing any of this exists.

Palette.glsl now converts with nothing left over, is compiled as a port
by Tests/Runtime, and is rendered by Tests/Runtime/ArrayTests - the third
stage whose result nothing on the CPU can observe. An array read at an
index that never varies is a perfectly plausible flat colour, so the
frame is what says which element each pixel got: a four-colour palette
subscripted by the pixel's own quarter comes out in four bands, and the
same palette clamped over an index that really does go negative comes out
the first colour on the left if the index is signed and the last one if
it is not. Nothing short of the frame distinguishes those two.

Corpus/Lattice.glsl is new, so the coverage table keeps measuring
something: the vector half of the row this stage closed the scalar half
of - an ivec2 cell, a componentwise lessThan, and the bvec2 it yields
with no any()/all() to collapse it.

One bug found while measuring that, and a bad-shaped one: the transpiler
hung - did not fail, hung - on any shader with a top-level struct.
Recovery from a construct the parser cannot keep skips to a semicolon and
stops at a closing brace without consuming it, so the brace a struct body
leaves behind was reported, skipped to, and reported again forever. A
struct is now named as the capability it is and skipped whole, and
top-level recovery asserts that it moved.

Coverage over the corpus goes from 10 of 11 to 11 of 12 clean. Every
integer and array row is gone; what is left is the vectors of both.

+++ Lib/shadertoy/Runtime/Program.h
@@ -107,9 +107,10 @@ private:
 // eacp ShaderProgram: it walks the uniform members and then calls define(),
 // which reaches mainImage through the vtable.
 //
-// The uniforms keep Shadertoy's names, so a ported body reads the same as the
-// GLSL it came from. Two deviations, both because the EDSL has no integer value
-// type usable in float arithmetic yet: iFrame is a float, and iDate is absent.
+// The uniforms keep Shadertoy's names and Shadertoy's types, so a ported body
+// reads the same as the GLSL it came from - iFrame included, which is an int on
+// the page and an Int here. One deviation is left: iDate is absent, which is a
+// clock this runtime does not have rather than a type the EDSL is missing.
 //
 // Texture channels are not declared here. A port declares the ones it samples
 // and no others, because every declared texture becomes a binding the draw has
@@ -124,7 +125,7 @@ public:
     GPU::Uniform<GPU::Float3> iResolution; // viewport size in pixels, then 1
     GPU::Uniform<GPU::Float> iTime; // seconds since the shader started
     GPU::Uniform<GPU::Float> iTimeDelta; // seconds the previous frame took
-    GPU::Uniform<GPU::Float> iFrame; // frames drawn since the start
+    GPU::Uniform<GPU::Int> iFrame; // frames drawn since the start
     GPU::Uniform<GPU::Float4> iMouse; // xy = pointer, zw = click (see below)
 
     // Uploads the fullscreen triangle and builds the pipeline. sampleCount comes
